Cast to number before string
For sending number instead of enum-string.
https://docs.gitlab.com/ee/api/protected_branches.html#protect-repository-branches

+++ src/GitLabApiClient/Models/Branches/Requests/ProtectBranchRequest.cs
@@ -29,9 +29,9 @@ namespace GitLabApiClient.Models.Branches.Requests
             Guard.NotEmpty(name, nameof(name));
             Name = name;
 
-            PushAccessLevel = pushAccessLevel.ToString();
-            MergeAccessLevel = mergeAccessLevel.ToString();
-            UnprotectAccessLevel = unprotectAccessLevel.ToString();
+            PushAccessLevel = ((int)pushAccessLevel).ToString();
+            MergeAccessLevel = ((int)mergeAccessLevel).ToString();
+            UnprotectAccessLevel = ((int)unprotectAccessLevel).ToString();
         }
 
         /// <summary>
